@@ -1,106 +1,104 @@
 {
-    "$schema": "./node_modules/ng-packagr/package.schema.json",
-    "name": "@zxing/ngx-scanner",
-    "description": "High-performance Angular 2+ barcode scanner component based on ZXing.",
-    "version": "1.1.0",
-    "private": false,
-    "ngPackage": {
-        "lib": {
-            "entryFile": "src/public_api.ts"
-        },
-        "whitelistedNonPeerDependencies": [
-            "@zxing/library",
-            "rxjs"
-        ]
+  "$schema": "./node_modules/ng-packagr/package.schema.json",
+  "name": "@zxing/ngx-scanner",
+  "description": "High-performance Angular 2+ barcode scanner component based on ZXing.",
+  "version": "1.1.0",
+  "private": false,
+  "ngPackage": {
+    "lib": {
+      "entryFile": "src/public_api.ts"
     },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/zxing-js/ngx-scanner"
+    "whitelistedNonPeerDependencies": [
+      "@zxing/library",
+      "rxjs"
+    ]
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/zxing-js/ngx-scanner"
+  },
+  "authors": [
+    {
+      "name": "David Werth",
+      "email": "werth.david@gmail.com"
     },
-    "authors": [
-        {
-            "name": "David Werth",
-            "email": "werth.david@gmail.com"
-        },
-        {
-            "name": "Luiz Machado",
-            "email": "machado@odahcam.com"
-        }
-    ],
-    "keywords": [
-        "angular",
-        "zxing",
-        "qr-code",
-        "scanner"
-    ],
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/zxing-js/ngx-scanner/issues"
-    },
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "start:ssl": "ng serve --ssl --live-reload false",
-        "build": "ng build",
-        "build:prod": "ng build --prod",
-        "build:packagr": "ng-packagr -p package.json",
-        "test": "ng test --single-run",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "docs": "npm run docs:build",
-        "docs:build": "compodoc -p tsconfig.json -n zxing-scanner -d docs --hideGenerator",
-        "docs:serve": "npm run docs:build -- -s",
-        "docs:watch": "npm run docs:build -- -s -w",
-        "publish:dist": "cd ./dist && npm publish --access=public"
-    },
-    "dependencies": {
-        "@zxing/library": "^0.5.0",
-        "rxjs": "^5.5.10"
-    },
-    "peerDependencies": {
-        "@angular/common": ">= 2.0.0 < 6",
-        "@angular/core": ">= 2.0.0 < 6",
-        "@angular/forms": ">= 2.0.0 < 6",
-        "core-js": "~2.5.5",
-        "zone.js": "^0.8.26"
-    },
-    "devDependencies": {
-        "@angular-devkit/core": "^0.3.2",
-        "@angular/animations": "^5.2.10",
-        "@angular/cli": "^1.7.4",
-        "@angular/common": "^5.2.10",
-        "@angular/compiler": "^5.2.10",
-        "@angular/compiler-cli": "^5.2.10",
-        "@angular/core": "^5.2.10",
-        "@angular/forms": "^5.2.10",
-        "@angular/http": "^5.2.10",
-        "@angular/language-service": "^5.2.10",
-        "@angular/platform-browser": "^5.2.10",
-        "@angular/platform-browser-dynamic": "^5.2.10",
-        "@angular/router": "^5.2.10",
-        "@compodoc/compodoc": "^1.1.1",
-        "@types/jasmine": "^2.8.6",
-        "@types/jasminewd2": "~2.0.2",
-        "@types/node": "^8.10.9",
-        "codelyzer": "^4.3.0",
-        "jasmine-core": "^2.99.1",
-        "jasmine-spec-reporter": "^4.2.1",
-        "karma": "^2.0.2",
-        "karma-chrome-launcher": "^2.2.0",
-        "karma-cli": "^1.0.1",
-        "karma-coverage-istanbul-reporter": "^1.4.2",
-        "karma-jasmine": "^1.1.1",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "ng-packagr": "^2.4.2",
-        "npm-install-peers": "^1.2.1",
-        "protractor": "^5.3.1",
-        "ts-node": "^4.1.0",
-        "tslint": "^5.9.1",
-        "typescript": "~2.7.2"
-    },
-    "main": "dist/bundles/zxing-ngx-scanner.umd.js",
-    "module": "dist/esm5/zxing-ngx-scanner.js",
-    "es2015": "dist/esm2015/zxing-ngx-scanner.js",
-    "typings": "dist/zxing-ngx-scanner.d.ts",
-    "metadata": "dist/zxing-ngx-scanner.metadata.json"
+    {
+      "name": "Luiz Machado",
+      "email": "machado@odahcam.com"
+    }
+  ],
+  "keywords": [
+    "angular",
+    "zxing",
+    "qr-code",
+    "scanner"
+  ],
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/zxing-js/ngx-scanner/issues"
+  },
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "start:ssl": "ng serve --ssl --live-reload false",
+    "build": "ng build",
+    "build:prod": "ng build --prod",
+    "build:packagr": "ng-packagr -p package.json",
+    "test": "ng test --single-run",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "docs": "npm run docs:build",
+    "docs:build": "compodoc -p tsconfig.json -n zxing-scanner -d docs --hideGenerator",
+    "docs:serve": "npm run docs:build -- -s",
+    "docs:watch": "npm run docs:build -- -s -w",
+    "publish:dist": "cd ./dist && npm publish --access=public"
+  },
+  "dependencies": {
+    "@zxing/library": "^0.5.0",
+    "rxjs": "^6.2.0"
+  },
+  "peerDependencies": {
+    "core-js": "~2.5.5",
+    "zone.js": "^0.8.26"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.6.6",
+    "@angular-devkit/core": "^0.3.2",
+    "@angular/animations": "^6.0.3",
+    "@angular/cli": "^6.0.7",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@compodoc/compodoc": "^1.1.1",
+    "@types/jasmine": "^2.8.6",
+    "@types/jasminewd2": "~2.0.2",
+    "@types/node": "^8.10.9",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "^2.99.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^2.0.2",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^1.0.1",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "^1.1.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^2.4.2",
+    "npm-install-peers": "^1.2.1",
+    "protractor": "^5.3.1",
+    "ts-node": "^4.1.0",
+    "tslint": "^5.9.1",
+    "typescript": "~2.7.2"
+  },
+  "main": "dist/bundles/zxing-ngx-scanner.umd.js",
+  "module": "dist/esm5/zxing-ngx-scanner.js",
+  "es2015": "dist/esm2015/zxing-ngx-scanner.js",
+  "typings": "dist/zxing-ngx-scanner.d.ts",
+  "metadata": "dist/zxing-ngx-scanner.metadata.json"
 }
